+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - Error translations
+- `beforeValidate` and `afterValidate` events
+
+### Changed
+- Removed duplicated code
 
 ## 1.0.3 - 2018-07-07